server/pyproject.toml
[project]
name = "chrome-plus-server"
version = "2.0.0"
description = "Chrome Plus AI Assistant - Backend Server with WebSocket and Celery support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core FastAPI and ASGI
    "fastapi>=0.104.0,<0.116.0",
    "uvicorn==0.34.2",
    "starlette>=0.40.0,<0.42.0",

    # WebSocket support
    "websockets>=12.0",
    "python-socketio>=5.10.0",

    # Celery and Redis
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",

    # HTTP clients and networking
    "aiohttp==3.11.18",
    "httpx[socks]==0.28.1",
    "httpx-sse==0.4.0",
    "requests==2.32.3",

    # AI and ML
    "openai==1.78.0",
    "pydantic-ai[tavily]==0.3.3",

    # Data validation and serialization
    "pydantic==2.11.4",
    "typing-extensions>=4.11,<5",

    # Configuration and environment
    "python-dotenv==1.1.1",
    "pyyaml==6.0.2",

    # Utilities
    "click==8.1.8",
    "colorama==0.4.6",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "idna==3.10",
    "urllib3==2.4.0",

    # Development and testing
    "pytest>=8.4.1",

    # Legacy dependencies (to be reviewed)
    "aiosignal==1.3.2",
    "anyio==4.9.0",
    "appnope==0.1.4",
    "boto3==1.38.13",
    "botocore==1.38.13",
    "cachetools==5.5.2",
    "decorator==5.1.1",
    "fastavro==1.10.0",
    "filelock==3.18.0",
    "huggingface-hub==0.31.1",
    "importlib-metadata>=6.0",
    "lxml==5.4.0",
    "multidict==6.4.3",
    "numpy==2.2.5",
    "packaging==24.2",
    "platformdirs==4.3.8",
    "psutil==5.9.7",
    "pyasn1==0.6.1",
    "regex==2024.11.6",
    "requests-toolbelt==1.0.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sse-starlette==2.3.4",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "yarl==1.20.0",
    "zipp==3.18.2",
]
